modified sh script to check if files or directories exist before removing

+++ cmakeGen.sh
@@ -1,4 +1,10 @@
 #/bin/sh
-rm CMakeCache.txt
-rm -rf CMakeFiles
+if [ -e "CMakeCache.txt" ]
+then
+	rm CMakeCache.txt
+fi
+if [ -d "CMakeFiles" ]
+then
+	rm -rf CMakeFiles
+fi
 cmake -G"Eclipse CDT4 - Unix Makefiles" -D CMAKE_BUILD_TYPE=Debug .
